Enable using highlight.js instead of htmlize.el for code blocks
When using highlight plugin, wrap src blocks in the required <pre><code>
tags. Additionally, in that case, htmlize.el fontification of the code
is inhibitied, since it should be carried out by hightlight.js. Also, do
not include the plugin by default in order not to break existing
documents relying on htmlize.